feat: implementa ClienteJuridicoService

+++ src/main/java/com/gestaooserp/dev/repository/ClienteJuridicoRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.ClienteJuridico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteJuridicoRepository extends JpaRepository<ClienteJuridico,Integer> {
+public interface ClienteJuridicoRepository extends JpaRepository<ClienteJuridico,Long> {
 }
